move document and annotation types to documents and annotations modules, respectively
